// src/API.ts
export type Product = {
  __typename: "Product",
  id: string,
  name?: string | null,
  isSold?: string | null,
  price?: number | null,
  image?: string | null,
  platformID: string,
  genreID: string,
  createdAt: string,
  updatedAt: string,
};